Summer-Backend/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.4
colorama==0.4.4
Django==4.0.2
django-cors-headers==3.11.0
docx2python==2.0.3
filelock==3.6.0
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-cloud-translate==3.6.1
google-cloud-vision==2.6.3
google-crc32c==1.3.0
google-resumable-media==2.3.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.1
joblib==1.1.0
lxml==4.8.0
Markdown==3.3.6
multidict==6.0.2
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
proto-plus==1.19.9
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
sentencepiece==0.1.96
simplet5==0.1.4
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.11.5
torch==1.10.2
torchmetrics==0.7.2
tqdm==4.62.3
transformers==4.16.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.7.0
